tests/wpt/tests/device-bound-session-credentials/not-secure-connection.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>No DBSC if connection is HTTP</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="helper.js" type="module"></script>

<script type="module">
  import { expireCookie, waitForCookie, addCookieAndSessionCleanup, setupShardedServerState } from "./helper.js";

  promise_test(async t => {
    await setupShardedServerState();
    const expectedCookieAndValue = "auth_cookie=abcdef0123";
    const expectedCookieAndAttributes = `${expectedCookieAndValue};Domain=${get_host_info().ORIGINAL_HOST};Path=/device-bound-session-credentials`;
    addCookieAndSessionCleanup(t, expectedCookieAndAttributes);

    // Prompt starting a session, and wait until registration completes.
    const login_response = await fetch('login.py');
    assert_equals(login_response.status, 200);
    // For HTTP, this call will time out, because the cookie is never set.
    assert_false(await waitForCookie(expectedCookieAndValue));
  }, "Try to establish a session over HTTP");
</script>
